refactor: move logic into new path lib
