+++ lua/neotest-golang/parse.lua
@@ -1,4 +1,5 @@
---- Parse the output from the go test command.
+--- This file is centered around the parsing/processing of test execution output
+--- and assembling of the final results to hand back over to Neotest.
 
 local async = require("neotest.async")
