+++ lua/neotest-golang/path.lua
@@ -1,0 +1,39 @@
+--- Helpers around filepaths.
+
+local M = {}
+
+--- Find a file upwards in the directory tree and return its path, if found.
+--- @param filename string
+--- @param start_path string
+--- @return string | nil
+function M.find_file_upwards(filename, start_path)
+  local scan = require("plenary.scandir")
+  local cwd = vim.fn.getcwd()
+  local found_filepath = nil
+  while start_path ~= cwd do
+    local files = scan.scan_dir(
+      start_path,
+      { search_pattern = filename, hidden = true, depth = 1 }
+    )
+    if #files > 0 then
+      found_filepath = files[1]
+      break
+    end
+    start_path = vim.fn.fnamemodify(start_path, ":h") -- go up one directory
+  end
+
+  if found_filepath == nil then
+    -- check if filename exists in the current directory
+    local files = scan.scan_dir(
+      start_path,
+      { search_pattern = filename, hidden = true, depth = 1 }
+    )
+    if #files > 0 then
+      found_filepath = files[1]
+    end
+  end
+
+  return found_filepath
+end
+
+return M
