+++ lua/neotest-golang/runspec_file.lua
@@ -2,7 +2,7 @@
 
 local cmd = require("neotest-golang.cmd")
 local convert = require("neotest-golang.convert")
-local runspec_dir = require("neotest-golang.runspec_dir")
+local path = require("neotest-golang.path")
 
 local M = {}
 
@@ -15,7 +15,7 @@ function M.build(pos, tree)
     return M.fail_fast(pos)
   end
 
-  local go_mod_filepath = runspec_dir.find_file_upwards("go.mod", pos.path)
+  local go_mod_filepath = path.find_file_upwards("go.mod", pos.path)
   if go_mod_filepath == nil then
     -- if no go.mod file was found up the directory tree, until reaching $CWD,
     -- then we cannot determine the Go project root.
